refactor: extract manifest history columns to stay under 300-line limit
The architecture size test enforces a 300-line limit on .tsx files.
Move buildColumns and its dependencies to manifest-history-columns.tsx
to bring the table component from 320 to 229 lines.
